apps/image-filter/src/requirements.txt
gradio==4.25.0
pillow
numpy
pandas
requests